organize login.js
